+++ src/cspark/sdk/_validators.py
@@ -1,5 +1,6 @@
 import re
 from abc import ABC, abstractmethod
+from json import loads
 from typing import List, Optional, Union
 
 from ._errors import SparkError
@@ -72,7 +73,32 @@ class BaseUrlValidator(BaseValidator):
             raise SparkError.sdk('must be a Spark base URL <*.coherent.global>', value)
 
 
+class TransformValidator(BaseValidator):
+    def __new__(cls):
+        if not hasattr(cls, 'instance'):
+            cls.instance = super(TransformValidator, cls).__new__(cls)
+        cls.instance.reset()
+        return cls.instance
+
+    def validate(self, value: str):
+        if not value:
+            raise SparkError.sdk('must be a properly defined JSON object', value)
+
+        try:
+            json_data = loads(value)
+        except Exception as cause:
+            raise SparkError.sdk(f'<{value}> must be a valid JSON object', cause) from cause
+
+        if not json_data.get('transform_type'):
+            raise SparkError.sdk("JSON object must have a 'transform_type' property", value)
+
+        target_version = json_data.get('target_api_version')
+        if not target_version or target_version not in ['v3', 'v4']:
+            raise SparkError.sdk("'target_api_version' must be either 'v3' or 'v4'", value)
+
+
 class Validators:
     empty_str = EmptyStringValidator
     positive_num = PositiveNumberValidator
     base_url = BaseUrlValidator
+    transform = TransformValidator